Reorganise view code into different modules.
